Refactor code with new alternatives
Explored and implemented alternative solutions to enhance functionality and address existing issues in the codebase.
[skip gpt_engineer]
